% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demplot.R
\name{demplot}
\alias{demplot}
\title{demplot function}
\usage{
demplot()
}
\value{
plot of these two variables
}
\description{
This function plots the "feeling thermometers" of survey respondents in regards
to the 2016 Democratic presidential candidate and each respondent's 2016
Democratic House of Representatives candidate against each other.
}
